renames ThemeProvider to NDSProvider
